jieyuzhang

+++ project/function.h
@@ -22,7 +22,7 @@ void free_list(Station_list *head);
 //ºËÐÄº¯Êý
 //cav
 void cav_init();
-void cav_update(int time);
+void cav_update(float time);
 void print_init();
 void cav_setstatus(Cavalier *cav);
 //alloc
